requirements.txt
Flask==1.1.1
pandas==1.0.1
Pillow==7.0.0
python-docx==0.8.10
requests==2.18.4